+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 agp = "9.0.0"
 kotlin = "2.3.20"
-lib-version-name = "1.0.2-alpha01"
+lib-version-name = "1.0.2-alpha02"
 
 # plugin
 compose-plugin = "1.10.3"
