+++ hig-core/src/iosMain/kotlin/zone/ien/hig/SystemBarAppearance.ios.kt
@@ -20,7 +20,7 @@ package zone.ien.hig
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
-import androidx.compose.ui.interop.LocalUIViewController
+import androidx.compose.ui.uikit.LocalUIViewController
 import org.jetbrains.skiko.OS
 import org.jetbrains.skiko.OSVersion
 import org.jetbrains.skiko.available
